修改整体架构

+++ User/inc/pid.h
@@ -22,7 +22,7 @@ typedef struct
 
 void PID_ParameterInit(PID* pid,float set_speed);
 
-float PID_SpeedAjust(PID* pid,u32 real_speed);
+float PID_SpeedAjust(PID* pid,float real_speed);
